test(review): split clean-room isolation coverage
