+++ crates/cli-sub-agent/src/review_convergence/clean_room_tests.rs
@@ -1,22 +1,16 @@
-use std::path::{Path, PathBuf};
+use std::path::Path;
 use std::sync::{Arc, Mutex};
 use std::time::Duration;
 
 use anyhow::{Result, anyhow};
-use csa_session::convergence::{
-    AdmittedModelIdentity, CommandAuthorityCatalogIdentity, CommandAuthorityPolicy,
-    CommandAuthoritySnapshot, CommandAuthoritySource, EpochRecord, GitObjectId, Sha256Digest,
-};
+use csa_session::convergence::{EpochRecord, GitObjectId, Sha256Digest};
 
 use super::clean_room::{
-    AdmittedProviderSessionFactory, CleanRoomWorkspaceFactory, CleanupFailureLedger,
-    DetachedWorkspaceDriver, DetachedWorkspacePlan, ExactOidWorkspaceFactory,
-    MaterializedWorkspace, ProviderSessionDriver, ProviderSessionFactory, ProviderSessionOutcome,
-    ProviderSessionRequest, WorkspaceCleanup,
+    CleanRoomWorkspaceFactory, CleanupFailureLedger, DetachedWorkspaceDriver,
+    DetachedWorkspacePlan, ExactOidWorkspaceFactory, MaterializedWorkspace, WorkspaceCleanup,
 };
-use crate::pipeline::{ParentSessionSource, SessionCreationMode};
 
-fn epoch() -> EpochRecord {
+pub(super) fn epoch() -> EpochRecord {
     EpochRecord::new(
         GitObjectId::parse(&"a".repeat(40)).expect("base oid"),
         GitObjectId::parse(&"b".repeat(40)).expect("head oid"),
@@ -24,23 +18,8 @@ fn epoch() -> EpochRecord {
     )
 }
 
-fn authority(reasoning: [&str; 2]) -> CommandAuthoritySnapshot {
-    CommandAuthoritySnapshot::new(
-        CommandAuthoritySource::tier("review", "test").expect("source"),
-        CommandAuthorityPolicy::new(false, vec!["codex".to_string()], false, true).expect("policy"),
-        CommandAuthorityCatalogIdentity::new("test catalog", "v1").expect("catalog"),
-        vec![
-            AdmittedModelIdentity::new("codex", "openai", "gpt-5.4", reasoning[0])
-                .expect("strongest identity"),
-            AdmittedModelIdentity::new("codex", "openai", "gpt-5.3", reasoning[1])
-                .expect("secondary identity"),
-        ],
-    )
-    .expect("authority")
-}
-
 #[derive(Clone)]
-struct FakeCleanup {
+pub(super) struct FakeCleanup {
     calls: Arc<Mutex<Vec<Duration>>>,
     fail: bool,
 }
@@ -55,7 +34,7 @@ impl WorkspaceCleanup for FakeCleanup {
     }
 }
 
-struct RecordingWorkspaceDriver {
+pub(super) struct RecordingWorkspaceDriver {
     plans: Arc<Mutex<Vec<DetachedWorkspacePlan>>>,
     cleanup_calls: Arc<Mutex<Vec<Duration>>>,
     observed_head: String,
@@ -80,14 +59,14 @@ impl DetachedWorkspaceDriver for RecordingWorkspaceDriver {
     }
 }
 
-type WorkspaceFactoryFixture = (
+pub(super) type WorkspaceFactoryFixture = (
     ExactOidWorkspaceFactory<RecordingWorkspaceDriver>,
     Arc<Mutex<Vec<DetachedWorkspacePlan>>>,
     Arc<Mutex<Vec<Duration>>>,
     CleanupFailureLedger,
 );
 
-fn factory(observed_head: String, cleanup_fails: bool) -> WorkspaceFactoryFixture {
+pub(super) fn factory(observed_head: String, cleanup_fails: bool) -> WorkspaceFactoryFixture {
     let plans = Arc::new(Mutex::new(Vec::new()));
     let cleanup_calls = Arc::new(Mutex::new(Vec::new()));
     let ledger = CleanupFailureLedger::default();
@@ -248,133 +227,6 @@ fn observed_head_mismatch_fails_closed_and_cleans_partial_workspace() {
         .expect_err("mismatched materialization must fail");
     assert!(error.to_string().contains("exact frozen head"));
     assert_eq!(cleanup_calls.lock().expect("cleanup").len(), 1);
-}
-
-#[test]
-fn provider_request_selects_strongest_xhigh_identity_and_proves_clean_room_policy() {
-    let temp = tempfile::tempdir().expect("tempdir");
-    let frozen = epoch();
-    let (mut factory, _plans, _cleanup_calls, _ledger) =
-        factory(frozen.head_oid().as_str().to_string(), false);
-    let guard = factory
-        .create(
-            &temp.path().join("source"),
-            &temp.path().join("room"),
-            &temp.path().join("bundle"),
-            frozen,
-        )
-        .expect("guard");
-
-    let request =
-        ProviderSessionRequest::from_authority(guard.workspace(), &authority(["xhigh", "high"]))
-            .expect("provider request");
-
-    assert_eq!(request.selected_model().model(), "gpt-5.4");
-    assert_eq!(request.selected_model().reasoning(), "xhigh");
-    assert_eq!(request.cwd(), guard.workspace().root());
-    assert!(request.readonly_project_root());
-    assert!(request.extra_writable().is_empty());
-    assert_eq!(
-        request.extra_readable(),
-        vec![guard.workspace().bundle_path().to_path_buf()]
-    );
-    assert_eq!(
-        request.parent_session_source(),
-        ParentSessionSource::ExplicitOnly
-    );
-    assert_eq!(
-        request.session_creation_mode(),
-        SessionCreationMode::FreshChild
-    );
-    assert!(request.startup_env().to_child_env_vars().is_empty());
-    assert!(request.parent().is_none());
-    assert!(request.resume_session().is_none());
-}
-
-#[test]
-fn provider_request_rejects_non_xhigh_strongest_identity() {
-    let temp = tempfile::tempdir().expect("tempdir");
-    let frozen = epoch();
-    let (mut factory, _plans, _cleanup_calls, _ledger) =
-        factory(frozen.head_oid().as_str().to_string(), false);
-    let guard = factory
-        .create(
-            &temp.path().join("source"),
-            &temp.path().join("room"),
-            &temp.path().join("bundle"),
-            frozen,
-        )
-        .expect("guard");
-    let error =
-        ProviderSessionRequest::from_authority(guard.workspace(), &authority(["high", "xhigh"]))
-            .expect_err("strongest identity must be xhigh");
-    assert!(error.to_string().contains("xhigh"));
-}
-
-struct NeverProviderDriver;
-
-impl ProviderSessionDriver for NeverProviderDriver {
-    fn run(
-        &mut self,
-        _admitted: &crate::pipeline::AdmittedExecutor,
-        request: &ProviderSessionRequest,
-    ) -> Result<ProviderSessionOutcome> {
-        panic!(
-            "provider driver must not execute in Slice 2 tests: {}",
-            request.cwd().display()
-        );
-    }
-}
-
-#[test]
-fn provider_adapter_type_is_bound_to_existing_admitted_executor_without_live_execution() {
-    fn assert_factory<T: ProviderSessionFactory>() {}
-    assert_factory::<AdmittedProviderSessionFactory<'static, NeverProviderDriver>>();
-}
-
-struct RecordingProviderFactory {
-    calls: Vec<PathBuf>,
-}
-
-impl ProviderSessionFactory for RecordingProviderFactory {
-    fn run(&mut self, request: &ProviderSessionRequest) -> Result<ProviderSessionOutcome> {
-        self.calls.push(request.cwd().to_path_buf());
-        Ok(ProviderSessionOutcome::new(
-            "01KCLEANROOMSESSION",
-            b"provider artifact",
-        ))
-    }
-}
-
-#[test]
-fn provider_port_is_mechanically_tested_with_a_fake_only() {
-    let temp = tempfile::tempdir().expect("tempdir");
-    let frozen = epoch();
-    let (mut workspace_factory, _plans, _cleanup_calls, _ledger) =
-        factory(frozen.head_oid().as_str().to_string(), false);
-    let guard = workspace_factory
-        .create(
-            &temp.path().join("source"),
-            &temp.path().join("room"),
-            &temp.path().join("bundle"),
-            frozen,
-        )
-        .expect("guard");
-    let request =
-        ProviderSessionRequest::from_authority(guard.workspace(), &authority(["xhigh", "high"]))
-            .expect("request");
-    let mut fake = RecordingProviderFactory { calls: Vec::new() };
-
-    let outcome = fake.run(&request).expect("fake provider outcome");
-
-    assert_eq!(fake.calls, vec![guard.workspace().root().to_path_buf()]);
-    assert_eq!(outcome.session_id(), "01KCLEANROOMSESSION");
-    assert_eq!(outcome.artifact(), b"provider artifact");
-    assert_eq!(
-        outcome.artifact_digest(),
-        &Sha256Digest::compute(b"provider artifact")
-    );
-    assert_eq!(request.epoch(), guard.workspace().epoch());
 }
 
 #[test]
